refactor(sanity): simplify client config resolution, make apiHost aware
